requirements.txt
asttokens==2.2.1
backcall==0.2.0
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
colorama==0.4.6
decorator==5.1.1
executing==1.2.0
google-api-core==2.11.0
google-auth==2.17.2
google-cloud-speech==2.19.0
googleapis-common-protos==1.59.0
grpcio==1.53.0
grpcio-status==1.53.0
idna==3.4
ipython==8.12.0
jedi==0.18.2
matplotlib-inline==0.1.6
parso==0.8.3
pickleshare==0.7.5
prompt-toolkit==3.0.38
proto-plus==1.22.2
protobuf==4.22.1
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.13
pydub==0.25.1
Pygments==2.14.0
requests==2.28.2
rsa==4.9
six==1.16.0
SpeechRecognition==3.10.0
stack-data==0.6.2
traitlets==5.9.0
typing_extensions==4.5.0
urllib3==1.26.15
wcwidth==0.2.6
